requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2
django-ninja==1.3.0
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
PyQRCode==1.2.1
python-decouple==3.8
qrcode==8.1
# scalar-django-ninja @ git+https://github.com/scalar/scalar.git@731b037cd05a545b9a0823ef5dd4f3d5bd28e8b1#subdirectory=packages/scalar_django_ninja
sqlparse==0.5.3
typing_extensions==4.12.2
